CIRCLE: added a cast in dxf_circle_write().

+++ src/circle.c
@@ -575,7 +575,7 @@ dxf_circle_write
         }
         if (circle->paperspace == DXF_PAPERSPACE)
         {
-                fprintf (fp->fp, " 67\n%hd\n", DXF_PAPERSPACE);
+                fprintf (fp->fp, " 67\n%hd\n", (int16_t) DXF_PAPERSPACE);
         }
         fprintf (fp->fp, "  8\n%s\n", circle->layer);
         if (strcmp (circle->linetype, DXF_DEFAULT_LINETYPE) != 0)
